+++ src/main/java/maroonshaded/gildedarmor/GildedArmor.java
@@ -1,6 +1,10 @@
 package maroonshaded.gildedarmor;
 
 import maroonshaded.gildedarmor.init.ModItems;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.tags.ItemTags;
+import net.minecraft.tags.TagKey;
+import net.minecraft.world.item.Item;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.common.Mod;
@@ -11,6 +15,8 @@ public class GildedArmor
 {
     public static final String MODID = "gildedarmor";
 
+    public static final TagKey<Item> ENDERITE_INGOT_TAG = ItemTags.create(location("enderite_ingot"));
+
     public GildedArmor()
     {
         IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
@@ -18,5 +24,10 @@ public class GildedArmor
         ModItems.ITEMS.register(modEventBus);
 
         MinecraftForge.EVENT_BUS.register(this);
+    }
+
+    public static ResourceLocation location(String path)
+    {
+        return new ResourceLocation(MODID, path);
     }
 }