+++ src/main/java/maroonshaded/gildedarmor/event/CreativeModeTabEventHandler.java
@@ -12,7 +12,6 @@ import net.minecraftforge.event.CreativeModeTabEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 
-@SuppressWarnings("unused")
 @Mod.EventBusSubscriber(modid = GildedArmor.MODID, bus = Mod.EventBusSubscriber.Bus.MOD)
 public class CreativeModeTabEventHandler
 {