+++ src/main/java/maroonshaded/gildedarmor/item/GildedArmorItem.java
@@ -1,6 +1,5 @@
 package maroonshaded.gildedarmor.item;
 
-import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ArmorItem;
 import net.minecraft.world.item.ArmorMaterial;
@@ -8,7 +7,7 @@ import net.minecraft.world.item.ItemStack;
 
 public class GildedArmorItem extends ArmorItem
 {
-    public GildedArmorItem(ArmorMaterial material, EquipmentSlot slot, Properties properties)
+    public GildedArmorItem(ArmorMaterial material, ArmorItem.Type slot, Properties properties)
     {
         super(material, slot, properties);
     }