+++ src/main/java/maroonshaded/gildedarmor/item/GildedEnderiteHelmetItem.java
@@ -1,6 +1,5 @@
 package maroonshaded.gildedarmor.item;
 
-import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.monster.EnderMan;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
@@ -9,7 +8,7 @@ public class GildedEnderiteHelmetItem extends GildedArmorItem
 {
     public GildedEnderiteHelmetItem(Properties properties)
     {
-        super(ModArmorMaterials.GILDED_ENDERITE, EquipmentSlot.HEAD, properties);
+        super(ModArmorMaterials.GILDED_ENDERITE, Type.HELMET, properties);
     }
 
     @Override