+++ src/main/java/maroonshaded/gildedarmor/item/ModArmorMaterials.java
@@ -2,41 +2,67 @@ package maroonshaded.gildedarmor.item;
 
 import com.google.common.base.Suppliers;
 import maroonshaded.gildedarmor.GildedArmor;
-import net.minecraft.resources.ResourceLocation;
+import net.minecraft.Util;
 import net.minecraft.sounds.SoundEvent;
 import net.minecraft.sounds.SoundEvents;
-import net.minecraft.tags.ItemTags;
-import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.util.StringRepresentable;
+import net.minecraft.world.item.ArmorItem;
 import net.minecraft.world.item.ArmorMaterial;
-import net.minecraft.world.item.ArmorMaterials;
+import net.minecraft.world.item.Items;
 import net.minecraft.world.item.crafting.Ingredient;
 
+import java.util.EnumMap;
 import java.util.function.Supplier;
-import java.util.function.ToIntFunction;
 
-public enum ModArmorMaterials implements ArmorMaterial
+public enum ModArmorMaterials implements StringRepresentable, ArmorMaterial
 {
-    GILDED_NETHERITE("gilded_netherite", ArmorMaterials.NETHERITE),
-    GILDED_ENDERITE("gilded_enderite", 8, new int[]{4, 7, 9, 4}, 17, SoundEvents.ARMOR_EQUIP_NETHERITE, 4.0F, 0.1F,
-            () -> Ingredient.of(ItemTags.create(new ResourceLocation(GildedArmor.MODID, "enderite_ingot"))),
-            true);
+    GILDED_NETHERITE("gilded_netherite", 37, Util.make(new EnumMap<>(ArmorItem.Type.class), map ->
+    {
+        map.put(ArmorItem.Type.BOOTS, 3);
+        map.put(ArmorItem.Type.LEGGINGS, 6);
+        map.put(ArmorItem.Type.CHESTPLATE, 8);
+        map.put(ArmorItem.Type.HELMET, 3);
+    }), 17, SoundEvents.ARMOR_EQUIP_NETHERITE, 3.0F, 0.1F,
+            () -> Ingredient.of(Items.NETHERITE_INGOT), true),
+    GILDED_ENDERITE("gilded_enderite", 8, Util.make(new EnumMap<>(ArmorItem.Type.class), map ->
+    {
+        map.put(ArmorItem.Type.BOOTS, 4);
+        map.put(ArmorItem.Type.LEGGINGS, 7);
+        map.put(ArmorItem.Type.CHESTPLATE, 9);
+        map.put(ArmorItem.Type.HELMET, 4);
+    }), 17, SoundEvents.ARMOR_EQUIP_NETHERITE, 4.0F, 0.1F,
+            () -> Ingredient.of(GildedArmor.ENDERITE_INGOT_TAG), true);
 
-    private static final int[] HEALTH_PER_SLOT = {13, 15, 16, 11};
-    private static final int[] ENDERITE_HEALTH_PER_SLOT = {128, 144, 160, 112};
+    private static final EnumMap<ArmorItem.Type, Integer> HEALTH_FUNCTION_FOR_TYPE = Util.make(new EnumMap<>(ArmorItem.Type.class), map ->
+    {
+        map.put(ArmorItem.Type.BOOTS, 13);
+        map.put(ArmorItem.Type.LEGGINGS, 15);
+        map.put(ArmorItem.Type.CHESTPLATE, 16);
+        map.put(ArmorItem.Type.HELMET, 11);
+    });
+    private static final EnumMap<ArmorItem.Type, Integer> ENDERITE_HEALTH_FUNCTION_FOR_TYPE = Util.make(new EnumMap<>(ArmorItem.Type.class), map ->
+    {
+        map.put(ArmorItem.Type.BOOTS, 128);
+        map.put(ArmorItem.Type.LEGGINGS, 144);
+        map.put(ArmorItem.Type.CHESTPLATE, 160);
+        map.put(ArmorItem.Type.HELMET, 112);
+    });
     private final String name;
-    private final ToIntFunction<EquipmentSlot> durability;
-    private final ToIntFunction<EquipmentSlot> protection;
+    private final int durabilityMultiplier;
+    private final boolean useEnderiteDurability;
+    private final EnumMap<ArmorItem.Type, Integer> protectionFunctionForType;
     private final int enchantmentValue;
     private final SoundEvent sound;
     private final float toughness;
     private final float knockbackResistance;
     private final Supplier<Ingredient> repairIngredient;
 
-    ModArmorMaterials(String name, int durabilityMultiplier, int[] damageReductionAmountArray, int enchantmentValue, SoundEvent sound, float toughness, float knockbackResistance, Supplier<Ingredient> repairIngredient, boolean useEnderiteDurability)
+    ModArmorMaterials(String name, int durabilityMultiplier, EnumMap<ArmorItem.Type, Integer> protectionFunctionForType, int enchantmentValue, SoundEvent sound, float toughness, float knockbackResistance, Supplier<Ingredient> repairIngredient, boolean useEnderiteDurability)
     {
         this.name = name;
-        durability = slot -> (useEnderiteDurability ? getEnderiteHealthPerSlot() : getHealthPerSlot())[slot.getIndex()] * durabilityMultiplier;
-        protection = slot -> damageReductionAmountArray[slot.getIndex()];
+        this.durabilityMultiplier = durabilityMultiplier;
+        this.useEnderiteDurability = useEnderiteDurability;
+        this.protectionFunctionForType = protectionFunctionForType;
         this.enchantmentValue = enchantmentValue;
         this.sound = sound;
         this.toughness = toughness;
@@ -44,38 +70,19 @@ public enum ModArmorMaterials implements ArmorMaterial
         this.repairIngredient = Suppliers.memoize(repairIngredient::get);
     }
 
-    ModArmorMaterials(String name, ArmorMaterial reference)
+    @Override
+    public int getDurabilityForType(ArmorItem.Type pType)
     {
-        this.name = name;
-        durability = reference::getDurabilityForSlot;
-        protection = reference::getDefenseForSlot;
-        enchantmentValue = reference.getEnchantmentValue();
-        sound = reference.getEquipSound();
-        toughness = reference.getToughness();
-        knockbackResistance = reference.getKnockbackResistance();
-        repairIngredient = reference::getRepairIngredient;
-    }
-
-    public static int[] getHealthPerSlot()
-    {
-        return HEALTH_PER_SLOT;
-    }
-
-    public static int[] getEnderiteHealthPerSlot()
-    {
-        return ENDERITE_HEALTH_PER_SLOT;
+        return (useEnderiteDurability
+                ? ENDERITE_HEALTH_FUNCTION_FOR_TYPE
+                : HEALTH_FUNCTION_FOR_TYPE)
+                .get(pType) * durabilityMultiplier;
     }
 
     @Override
-    public int getDurabilityForSlot(EquipmentSlot pSlot)
+    public int getDefenseForType(ArmorItem.Type pType)
     {
-        return durability.applyAsInt(pSlot);
-    }
-
-    @Override
-    public int getDefenseForSlot(EquipmentSlot pSlot)
-    {
-        return protection.applyAsInt(pSlot);
+        return protectionFunctionForType.get(pType);
     }
 
     @Override
@@ -112,5 +119,11 @@ public enum ModArmorMaterials implements ArmorMaterial
     public float getKnockbackResistance()
     {
         return knockbackResistance;
+    }
+
+    @Override
+    public String getSerializedName()
+    {
+        return getName();
     }
 }