+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Gilding upgrade smithing template (1.20 experimental feature).
+- Gilded armor can be trimmed.
+
+### Changed
+- Updated to 1.19.4
+
 ## [1.5.1] - 2023-04-15
 ### Added
 - Russian translation for the gilded enderite armor.